Added Tuple Support, Prefabs, and Baseline PubSub
Tuple support is not mine but is public domain code:
https://gist.github.com/michaelbartnett/5652076

Adapted the singleton pattern for MonoBehaviour.

Added prefabs for the background plane, player, lighting, and camera.

Attempted an Input Manager, but decided it was not worth the effort
considering the use it was going to get. Too much high level access from
unity. Need lower level functionality, and UnityEngine.Event was not
cutting it. Best to use the default input.