The Motherlode
So many things where to start?

Added a second player prefab. Separate controls not implemented. But
wait there's more

Organized the prefabs folder better and renamed many prefabs to better
match their function.

'PlayerUI.cs' has been renamed...to 'UnitNameplate.cs. Why? The entire
class has been reworked to allow any 'GameObject' which implements
'IStats' to use the class. When enemies are added, just instantiate a
nameplate from a custom prefab for enemies, and you have a nameplate.

The UI for skills has been reworked slightly. It now accounts for all
player objects currently in the scene and will display all of them that
it can centered along the bottom of the screen. It's not perfect, but
still pretty nice. This also means the following:

The UI for skills uses it's own object from 'SkillButton.cs'.
Technically there is nothing stopping you from binding this object to
any class which implements the 'IUsesSkills' interface. Once an enemy is
added, all you have to do is create an instance of this class, set the
'parent' and 'skillIndex' variable and the class will do the rest.

Some changes to organization and naming of Interface items

Added tags for each menu item. Will use these later, did enough tonight
already.