Got Web Builds to work
Required a preprocessor directive of 'UNITY_WEBGL' to make sure to not
include any XInput code when building for web. Sucks, because Unity's
game pad support is terrible, but it's better than not having it at all.
Standalone builds and all other builds that work with .dlls will still
use XInput since it is the clearly superior option.