XML Config File For KeyBindings
Started work on making different Controller types such as
'UserController'. This will enable any object which implements the
'IControllable' interface to be controlled by any of the different
controller classes which implement 'IController'. This will in the end
allow the player to control enemies or for the player to have AI
patterns. Why? I don't, why not.

XML File is partially configured. Need help from mathew at this point.
Everything is working, but only just so. Don't understand why some of
the XML reading is the way it is or how to format the XML file properly.

Lots of other changes to make an 'IChildable' interface and
'IParentable' class which ensures that an object can be the parent of
another object or become the child of another object. Helps for code
clarity, but could be accomplished without it by just passing in
GameObjects. That's no fun though.

Currently, the two players in the scene are controlled by wasd and arrow
keys separately.