The Control Commit
Updated how user input is parsed. Allowing scalability as to how each
unit is controlled by creating 'IController' implementing classes to
handle controlling a unit. This should allow a unit to be controlled by
the user regardless of if it is a 'Player' unit or an 'Enemy' unit or
whatever, just so long as that class implements the 'IControllable'
interface. So far, you are able to easily and seamlessly control a
player an enemy unit. For the enemy though, at the moment he is also
being controlled by his 'NavMeshAgent' but that will change later.

Added a class to create, read, and write a 'KeyBindings.xml' file. This
already allows you to customize what the controls of each user is by
just editing the config file. It will get expanded upon later to include
more controls, but even for now it is functional. Try it out.

I deleted the second player prefab. I don't think we will need it. The
'GameManager' will eventually instantiate player prefabs based on a user
selection and color the material differently. Shouldn't need a prefab
for each player.