The Skill Commit
Changed the way skills are stored in data. This way should be much more
scalable. The prefab you instantiate when using a skill no longer knows
everything about the skill, only relevant information.

Changed how the UIManager builds the skill indicators to include any
object that implements 'IUsesSkills' and uses the 'ControlType' of
'UserControl'. This way eventually if the player is in control of the
Fortress or an enemy the UI can rebuild itself and display those skills.

Changed folder structure to be able to include a base class of 'Unit'
and 'Skill' namespace.

Oh yeah, and now units can take damage and potentially die.