The Base UIAnnouncer Commit
Added functionality to display a message on the screen and animate it
being revealed. It's put into a queue so multiple messages can be added
to the queue and displayed only once another one has finished animating.

The player announces it's name to the 'UIAnnouncer' as an example of the
functionality.