Experimental Changes to the Announcer
Some experimental changes to the Announces to allow for any amount of
animations to be done to the text when announcing a message.

Most of the framework is set up, but it's not as intuitive as I would
like. Some changes need to be made in order to make it easier to
use/scale.