Messy Commit for UIAnnouncer
Lots of work done, but not a lot to show for it. Started work on a base
Animations static class. Works pretty well, but took forever. Is not
very generic currently.

Started work on a custom property drawer for 'AnimationSequence'.
Learning how to parse a serialized property from Unity.

Small changes to prefabs