The Item Drops and Experience Value Exchange Commit
Enemies now have preset experience to give the player that kills them.
This is done in the Fireball Script. These values could change later
once we discuss as a group what the values should be.
Enemies are all now tagged as "Enemy" so upon death when a unit dies, it
can be tag checked if its an Enemy and if it is then the player can
receive experience.
Serialized the experience stat field in unit class so we can assign the
values appropriately for each enemy respectively.
Added 2 Properties to Item Drops class, one for health and one for mana.
Just assign the prefab a number to the respective stat you want the
object to increase.
Players can now receive health and mana from drops.