The Damage Text Commit
Added floating damage text. Uses a world space canvas, so it
unfortunatly cannot be seen above UI elements. I'm not sure of a work
around for this.

Changed some trigger/collision checkboxes to try to get better collision
resolution results, but to no avail. I think I may have to switch
movement to be based on rigidbody.AddForce, which I hate, but seems to
be the solution.