The Physics Commit
Changed the player to use Rigidbody.velocity. Collisions with objects
should work properly now as well as trigger fires.

Adjusted the 'UIAnnouncer' script to account for the new
'AnimationSequence' class. Works pretty nicely. F4 will send a test
announcement. Check it out.

Changed the namespace of the editor scripts to account for the fact that
there is a class in Unity called Editor...So these scripts should not be
in the 'Editor' namespace.