The Sick New Assets Commit
Changed the camera scripting and control scheme to fit a third person
style game. There is now only one player per computer since the mouse is
required, and split screen seems like a doable, but not worthy task.

Added prefabs for most of the new assets to work with the old set up.

The melee skill has been implemented and now does damage.

The mouse can be used to rotate the camera by holding the right mouse
button. You can also use the right analog stick on a controller for the
same effect.

So much math involved in making the player move based on the
camera...but it's there. Sloppy or not.

Added a background image to all of the nameplates to make them more
visible. More need to be added to the UIAnnouncer's log, and possibly an
outline needs to be added to the Announcements.

Broke Donte's options menu....

Messed around with how the item pickups are created. Added a rigid body
and gave them a random force. Not what is desired in a final build.