The First Vacation Commit(You must read this commit message)
You thought I wasn't going to work on this?

Added a third skill button to the default config
Added the target mode button to the config. Default key is 'F'

Changed the player script to not be a type of unit but rather the
user/player of the game. The player has a camera, a unit, and an
IController. This should allow even easier swapping between units for
the player to control.

Moved all mouse input for camera movement to the camera script.

Added a basic targeting system. Press F when facing a unit to lock on to
that enemy. The camera is then set to specific values and your character
strafes while looking at that unit. Not as easy to aim as I'd like it to
be

Rearranged the HUD and UI to better fit the targeting camera. This could
change though, as I'm not sure if I like it any better. A much better
solution would be to fade the health and mana bars out when the player's
position is behind them instead. We'll see.

Added a basic lightning skill. Works using Physics.SphereCast and
Physics.SphereOverlap. Not quite as easy to aim as I need it to be.

Fixed bugs with the instructions menu where it was not looping through
all skills the player has.

Changed skill buttons to display contextual hotkey information depending
on if a controller was plugged in when starting the game or not. Needs
to be updated later to display the correct information at all times.

Forced UnitNameplates to stop all coroutines upon being destroyed.

Added many more skill icons