Added Sounds and Audio Manager
Added Fireball sound and Melee Sound. When a unit uses the melee skill
then sound will play respectively. When a unit uses the fireball skill
then the sound will play respectively.  Created an Audio Manager class
that will be used to manage audio in the game.  The script is attached
to the game manager prefab. The number of sounds and clips to be played
will be set on the script in the inspector. The Next thing i will be
working on is finding a way for the user to alter the volumes of each
skill through the options menu.