Added Slider to options menu and worked on bugs
Added a volume slider to the options menu and now the user can change
the volume of the sound effects in the game. There was a bug in the game
that if the player would die then the game wouldnt go to the gameover
event. Added the Broadcast to that event inside the AIController script
in the OnUnitDied Function. Cleaned up the audio manager script a bit
and got rid of useless code. Added back ground music to the main menu
but it is still currently buggy and i will need to fix that