Destroying of itemdrops have been changed a bit
When a enemy dies a check is ran..if the players health is greater than
or equal to his maxHealth then the health item drop will be deleted
after 3 seconds. If the players mana is greater than or equal to his
maxMana then the mana item drop will be deleted after 3 seconds.
However, these items will not get deleted if the players health aand/or
mana or less than the max values, respectively.