The Strafe Commit
This is not even close to what I want the final animation to look like,
but this is a start and it proves that I understand how to manipulate
the blend tree of an animator.

The final animator needs to be given more values to decide what
animation to play. Currently, it too wild and values jump around for no
real reason. While the animations play decently for some cases, in other
cases it looks completely broken.