The Skill Commit
The skill script now has a static int m_maxLevel variable  that is set
to 10.
The max level a skill can be is 10.
I made a static public property to get the m_maxLevel variable. I made
these static due to the fact that there are places in code that an
instance is not needed and all we want to do is reference this
variable...so no need for an instance.
In the SkillButton script, in the OnCanUpgradeSkill function, I added a
line of code that will check if the current skills level is equal to or
greater than the max level variable and if so then just return out the
function. This allows the user to no longer be able to level up this
skill because it is at max and the upgrade button will no longer be
activated for this skill.
In the BaseSkill script, I added a line of code that will set the
current sprite based on an algorithm created in the GetSkillData
function.  In the Unit script, in the OnUpgradeSkill function, I added a
line of code to check if the skills current level is equal to the max
level and if its true then it will announce to the user the skills name
and that it is now maxed.